+++ frontend/packages/kubevirt-plugin/src/constants/vm/constants.ts
@@ -43,24 +43,6 @@ export enum DeviceType {
   DISK = 'DISK',
 }
 
-export enum VMWizardActionLabels {
-  YAML = 'New from YAML',
-  WIZARD = 'New with Wizard',
-  IMPORT = 'Import with Wizard',
-}
-
-export enum VMWizardName {
-  YAML = 'yaml',
-  WIZARD = 'wizard',
-  IMPORT = 'wizardImport',
-}
-
-export enum VMWizardMode {
-  IMPORT = 'import',
-  TEMPLATE = 'template',
-  VM = 'vm',
-}
-
 export const VM_DETAIL_EVENTS_HREF = 'events';
 
 export const PAUSED_VM_MODAL_MESSAGE =
