+++ frontend/packages/kubevirt-plugin/src/constants/vm/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './network';
 export * from './storage';
+export * from './ui';
