+++ frontend/packages/kubevirt-plugin/src/constants/vm/ui.ts
@@ -1,0 +1,17 @@
+export enum VMWizardActionLabels {
+  YAML = 'New from YAML',
+  WIZARD = 'New with Wizard',
+  IMPORT = 'Import with Wizard',
+}
+
+export enum VMWizardName {
+  YAML = 'yaml',
+  WIZARD = 'wizard',
+  IMPORT = 'wizardImport',
+}
+
+export enum VMWizardMode {
+  IMPORT = 'import',
+  TEMPLATE = 'template',
+  VM = 'vm',
+}
